apps/api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.1,<4.0
pydantic>=2.8,<3.0
pydantic-settings>=2.2,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
redis>=5.0,<7.0
email-validator>=2.0,<3.0
httpx>=0.27,<1.0
pytest>=8.3,<9.0
